Cleanup

+++ src/benchmark/SortInterface.java
@@ -5,7 +5,7 @@ public interface SortInterface {
     void recursiveSort(int[] list) throws UnsortedException;
     void iterativeSort(int[] list) throws UnsortedException;
 
-    int getCount();
+    long getCount();
     long getTime();
 
 }
